requirements.txt
fastai
cupy-cuda12x
numba
